refactor(scan): split 1813-line scan.py into 3 modules
Extract scan_preflight.py (100 lines) and scan_output.py (640 lines)
from scan.py (now 1105 lines). Re-exports in scan.py preserve all
existing imports for backwards compatibility.

- scan_preflight.py: _needs_ollama, _preflight_ollama_check, _ensure_scan_dependencies
- scan_output.py: _render_text_report, _display_llm_summary, _display_frame_cost_breakdown,
  _display_memory_stats, _render_contract_mode_summary, _generate_configured_reports,
  _write_ai_status_file, _update_tech_debt_file
- scan.py keeps: scan_command, _run_scan_async, _process_stream_events (closure-bound)
